Renamed api_to_model/ -> converters/ and api_wrapper/ -> fetchers/
