+++ backend/src/worker/fetchers/genres.py
@@ -1,21 +1,21 @@
-from src.worker.api_wrapper.paged_fetcher import PagedFetcher
+from src.worker.fetchers.paged_fetcher import PagedFetcher
 
 
-class HallFetcher(PagedFetcher):
+class GenreFetcher(PagedFetcher):
     """
-    This class stands in for fetching halls from the viernulvier API.
+    This class stands in for fetching genres from the viernulvier API.
 
     Its init takes in a `VNV_Wrapper` which will be used to execute all
-    requests, and has a `get_new_halls_after(timestamp)` to request all
-    halls.
+    requests, and has a `get_new_genres_after(timestamp)` to request all
+    genres.
     """
 
-    def get_new_halls_after(self, timestamp) -> list:
+    def get_new_genres_after(self, timestamp) -> list:
         """
-        Get all halls after the given timestamp.
+        Get all genres after the given timestamp.
 
         The timestamp is used **inclusive**, meaning that it probably returns
-        a hall that already exists in the database.
+        a genre that already exists in the database.
 
         When calling the API fails this will throw an error. However,
         if there was already data fetched (f.e. when paging and hitting a
@@ -24,8 +24,8 @@ class HallFetcher(PagedFetcher):
 
         ---
 
-        :param timestamp: used to get new halls after (inclusive)
+        :param timestamp: used to get new genres after (inclusive)
         """
 
         parameters = {"created_at[after]": timestamp}
-        return self.fetch_all("/halls", parameters)
+        return self.fetch_all("/genres", parameters)
