+++ backend/src/worker/fetchers/halls.py
@@ -1,20 +1,21 @@
-from src.worker.api_wrapper.paged_fetcher import PagedFetcher
+from src.worker.fetchers.paged_fetcher import PagedFetcher
 
 
-class TagFetcher(PagedFetcher):
+class HallFetcher(PagedFetcher):
     """
-    This class stands in for fetching tags from the viernulvier API.
+    This class stands in for fetching halls from the viernulvier API.
 
     Its init takes in a `VNV_Wrapper` which will be used to execute all
-    requests, and has a `get_new_tags_after(timestamp)` to request all tags.
+    requests, and has a `get_new_halls_after(timestamp)` to request all
+    halls.
     """
 
-    def get_new_tags_after(self, timestamp) -> list:
+    def get_new_halls_after(self, timestamp) -> list:
         """
-        Get all tags after the given timestamp.
+        Get all halls after the given timestamp.
 
         The timestamp is used **inclusive**, meaning that it probably returns
-        a tag that already exists in the database.
+        a hall that already exists in the database.
 
         When calling the API fails this will throw an error. However,
         if there was already data fetched (f.e. when paging and hitting a
@@ -23,8 +24,8 @@ class TagFetcher(PagedFetcher):
 
         ---
 
-        :param timestamp: used to get new tags after (inclusive)
+        :param timestamp: used to get new halls after (inclusive)
         """
 
         parameters = {"created_at[after]": timestamp}
-        return self.fetch_all("/tags", parameters)
+        return self.fetch_all("/halls", parameters)
