+++ backend/src/worker/fetchers/tags.py
@@ -1,21 +1,20 @@
-from src.worker.api_wrapper.paged_fetcher import PagedFetcher
+from src.worker.fetchers.paged_fetcher import PagedFetcher
 
 
-class GenreFetcher(PagedFetcher):
+class TagFetcher(PagedFetcher):
     """
-    This class stands in for fetching genres from the viernulvier API.
+    This class stands in for fetching tags from the viernulvier API.
 
     Its init takes in a `VNV_Wrapper` which will be used to execute all
-    requests, and has a `get_new_genres_after(timestamp)` to request all
-    genres.
+    requests, and has a `get_new_tags_after(timestamp)` to request all tags.
     """
 
-    def get_new_genres_after(self, timestamp) -> list:
+    def get_new_tags_after(self, timestamp) -> list:
         """
-        Get all genres after the given timestamp.
+        Get all tags after the given timestamp.
 
         The timestamp is used **inclusive**, meaning that it probably returns
-        a genre that already exists in the database.
+        a tag that already exists in the database.
 
         When calling the API fails this will throw an error. However,
         if there was already data fetched (f.e. when paging and hitting a
@@ -24,8 +23,8 @@ class GenreFetcher(PagedFetcher):
 
         ---
 
-        :param timestamp: used to get new genres after (inclusive)
+        :param timestamp: used to get new tags after (inclusive)
         """
 
         parameters = {"created_at[after]": timestamp}
-        return self.fetch_all("/genres", parameters)
+        return self.fetch_all("/tags", parameters)
